fix: 4 unrelated test failure clusters from WIP backlog
NotificationTemplate: add renderRaw() and split rendering paths.
render() now HTML-escapes context (safe for non-Blade callers);
renderRaw() preserves prior behavior for Blade templates that
already e()-escape. Fixes NotificationMailContentSafetyTest.

DatabaseAgnosticQueries: %u (ISO week) maps to MySQL %v, not %u.
MySQL's %u is week-of-year with Monday start but NOT ISO. Test
suite updated to assert against the corrected mapping plus a new
pgsql year SQL test (CAST(EXTRACT(YEAR FROM ...) AS INTEGER)).

intasend_transactions migration: net_amount, landlord_amount, and
intasend_invoice_id changed to nullable. They're populated post-
gateway-callback, not at intent creation, so requiring them at
INSERT broke 50+ webhook/payment tests with "Field has no default".

config/security.php: cast deletion_grace_days and data_retention_
years to (int). PHP 8.4 strict type coercion rejected string-from-
env on `int $retentionYears` in ArchiveOldPaymentsTest setUp().

+++ config/security.php
@@ -164,8 +164,8 @@ return [
     'compliance' => [
         'gdpr_enabled' => env('GDPR_ENABLED', true),
         'kenya_dpa_enabled' => env('KENYA_DPA_ENABLED', true),
-        'deletion_grace_days' => env('DELETION_GRACE_DAYS', 30),
-        'data_retention_years' => env('DATA_RETENTION_YEARS', 7),
+        'deletion_grace_days' => (int) env('DELETION_GRACE_DAYS', 30),
+        'data_retention_years' => (int) env('DATA_RETENTION_YEARS', 7),
         'consent_required' => ['terms', 'privacy'],
         'breach_notification_hours' => 72, // GDPR & Kenya DPA requirement
         'log_in_console' => env('AUDIT_LOG_IN_CONSOLE', false), // Log audit events in console/CLI
